requirements.txt
urllib3
pandas
python-dotenv
sqlalchemy
dask
pyodbc
pymysql
openpyxl
pyarrow
fastparquet
apache-airflow
apache-airflow-providers-google